not fully working

+++ src/strategy/raids/aq40/RaidAq40Actions.cpp
@@ -68,7 +68,7 @@ bool Aq40MoveFromOtherEmperorAction::Execute(Event event)
     if (Unit* boss2 = AI_VALUE2(Unit*, "find target", "emperor vek'nilash"))
     {
         ObjectGuid botguid = bot->GetGUID();
-        ObjectGuid petguid = (ObjectGuid)0UL;
+        ObjectGuid petguid = (ObjectGuid)(uint64)0UL;
         if (Pet* pet = bot->GetPet())
         {
             petguid = pet->GetGUID();
